+++ src/Core/Api/PaymentMethodConfiguration/Service/PaymentMethodConfigurationService.php
@@ -684,6 +684,13 @@ class PaymentMethodConfigurationService {
 			], $context);
 
 			// Media insert/update
+
+			// detect if collision, return existing id
+			$existingId = $this->checkMediaAlreadyExists($paymentMethodConfiguration->getResolvedImageUrl(), $context);
+			if ($existingId) {
+				return $existingId;
+			}
+
 			$mediaDefinition = $this->container->get(MediaDefinition::class);
 			$this->mediaSerializer->setRegistry($this->serializerRegistry);
 
@@ -702,6 +709,32 @@ class PaymentMethodConfigurationService {
 			$this->logger->critical($e->getMessage(), [$e->getTraceAsString()]);
 			return null;
 		}
+	}
+
+	/**
+	 * Check if the payment method icon is already uploaded.
+	 *
+	 * @param string                                                      $paymentMethodUrl
+	 * @param \Shopware\Core\Framework\Context                            $context
+	 *
+	 * @return string|bool
+	 */
+	private function checkMediaAlreadyExists($paymentMethodUrl, $context) {
+		// detect if collision, return existing id
+		if (!preg_match('#/([^/]+)\.[^/.]+$#', $paymentMethodUrl, $matches)) {
+			return false;
+		}
+		$filename = $matches[1];
+
+		$criteria = new Criteria();
+		$criteria->addFilter(new EqualsFilter('fileName', $filename));
+
+		$existing = $this->mediaRepository->search($criteria, $context)->first();
+
+		if ($existing) {
+			return $existing->getId();
+		}
+		return false;
 	}
 
 	/**