+++ src/Core/Api/WebHooks/Strategy/WebHookStrategyBase.php
@@ -378,16 +378,6 @@ abstract class WebHookStrategyBase implements WebHookStrategyInterface {
 	protected function unholdAndCancelDelivery(string $orderId, Context $context): void
 	{
 		$order = $this->getOrderEntity($orderId, $context);
-		try {
-			$this->orderService->orderStateTransition(
-				$order->getId(),
-				StateMachineTransitionActions::ACTION_CANCEL,
-				new ParameterBag(),
-				$context
-			);
-		} catch (\Exception $exception) {
-			$this->logger->info($exception->getMessage(), $exception->getTrace());
-		}
 
 		try {