+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 6.2.3
+- Fix issue with cart clearing on payment method change
+- Fixed issue with missing payment methods
+
 # 6.2.1
 - Fixed issue with multiple discount codes